refactor(web): keep the Loro envelope schema out of the loro-crdt module
Reading a document's content timestamp needs one zod schema. It lived in
`loro-store.ts`, which imports `loro-crdt` at module scope — so
`browser-local-store.ts` importing it put the whole CRDT library on the
critical path, for a string.

The bundle-size gate caught it, and it is the first time that gate has run in a
while: `verify` needs `test-browser`, which has been failing, so a skipped
`verify` had been hiding this.

Measured, gzip, critical-path JS:

| | files | size |
|---|---|---|
| main | 22 | 122.0 KB |
| this branch, schema imported from `loro-store` | 24 | **146.3 KB** (budget 126) |
| this branch, schema in its own module | 22 | 122.3 KB |

So the 24.3 KB was entirely `loro-crdt`; the feature itself costs 0.3 KB. My
own estimate before measuring was "a few hundred bytes", which was right about
the feature and wrong about the import — the reason to have the gate rather
than an opinion.

`loro-record-envelope.ts` is zod and nothing else, and its doc comment records
why it is not simply in the file that uses it most.
