+++ apps/web/src/lib/loro-store.ts
@@ -1,29 +1,7 @@
 import { Loro } from 'loro-crdt'
-import { z } from 'zod'
 import { openWhiteboardDb } from './browser-idb.js'
 import { shouldCompact } from './loro-compaction.js'
-
-/**
- * Versioned envelope for Loro records persisted in IndexedDB.
- * Single parse boundary — every IDB read for loroCanvases goes through this.
- * Type is derived via z.infer; no parallel hand-written interface.
- *
- * v: envelope format version (literal 1). Bump this when the envelope shape
- *    changes in a backward-incompatible way; old records will be rejected as
- *    'corrupt-snapshot' and the caller is responsible for recovery.
- */
-// z.custom pins the inferred type to Uint8Array<ArrayBuffer> (matching lib:ES2020+DOM),
-// which is narrower than the z.instanceof(Uint8Array) result (Uint8Array<ArrayBufferLike>).
-const uint8ArraySchema = z.custom<Uint8Array>((v) => v instanceof Uint8Array)
-
-export const loroRecordEnvelopeSchema = z.object({
-  v: z.literal(1),
-  snapshot: uint8ArraySchema,
-  updatedAt: z.string(),
-  deltas: z.array(uint8ArraySchema).optional(),
-})
-
-type LoroRecordEnvelope = z.infer<typeof loroRecordEnvelopeSchema>
+import { type LoroRecordEnvelope, loroRecordEnvelopeSchema } from './loro-record-envelope.js'
 
 export type LoroLoadResult =
   | { kind: 'ok'; snapshot: Uint8Array; deltas?: Uint8Array[] }
